Refactor: モデルのファクトリでの名前空間を修正し、テストクラスの名前空間を整理
