+++ app/Http/Requests/Manager/Tag/DeleteRequest.php
@@ -14,6 +14,7 @@ class DeleteRequest extends FormRequest
         return true;
     }
 
+    #[\Override]
     protected function prepareForValidation()
     {
         $this->merge([
