+++ database/factories/model/AttendanceFactory.php
@@ -2,17 +2,16 @@
 
 namespace Database\Factories\Model;
 
-use App\Model\Attendance;
 use App\Model\Course;
 use App\Model\Student;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Attendance>
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Model\Attendance>
  */
 class AttendanceFactory extends Factory
 {
-    protected $model = Attendance::class;
+    protected $model = \App\Model\Attendance::class;
 
     /**
      * Define the model's default state.
