+++ database/factories/model/InstructorFactory.php
@@ -6,7 +6,7 @@ use App\Model\Instructor;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Model>
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Model\Instructor>
  */
 class InstructorFactory extends Factory
 {
