+++ tests/Feature/Api/Instructor/Chapter/SortTest.php
@@ -11,6 +11,7 @@ class SortTest extends TestCase
     use RefreshDatabase;
 
     // setup
+    #[\Override]
     protected function setUp(): void
     {
         parent::setUp();
