+++ tests/Feature/Api/Manager/Tag/DeleteTest.php
@@ -10,6 +10,7 @@ class DeleteTest extends TestCase
 {
     use RefreshDatabase;
 
+    #[\Override]
     protected function setUp(): void
     {
         parent::setUp();
