+++ tests/Feature/Api/Student/Attendance/CompleteAllLessonsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Api\Student;
+namespace Tests\Feature\Api\Student\Attendance;
 
 use App\Model\Student;
 use Illuminate\Foundation\Testing\RefreshDatabase;
